# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.12.3 🍺
 * Do not edit manually.
 * Api
 * Email CRM API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export interface LoginInput {
  email: string;
  /** @minLength 1 */
  password: string;
}

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  admin: 'admin',
  user: 'user',
} as const;

export interface User {
  id: number;
  email: string;
  name: string;
  role: UserRole;
  /** @nullable */
  region?: string | null;
  isActive?: boolean;
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export type UserInputRole = typeof UserInputRole[keyof typeof UserInputRole];


export const UserInputRole = {
  admin: 'admin',
  user: 'user',
} as const;

export interface UserInput {
  email: string;
  name: string;
  /** @minLength 6 */
  password: string;
  role: UserInputRole;
  /** @nullable */
  region?: string | null;
}

export type UserUpdateRole = typeof UserUpdateRole[keyof typeof UserUpdateRole];


export const UserUpdateRole = {
  admin: 'admin',
  user: 'user',
} as const;

export interface UserUpdate {
  name?: string;
  role?: UserUpdateRole;
  /** @nullable */
  region?: string | null;
  isActive?: boolean;
  /** @minLength 6 */
  password?: string;
}

export interface Account {
  id: number;
  name: string;
  email: string;
  region: string;
  smtpHost: string;
  smtpPort: number;
  smtpUser?: string;
  smtpPass?: string;
  isActive: boolean;
  dailyLimit?: number;
  sentToday?: number;
  /** @nullable */
  signature?: string | null;
  createdAt: string;
}

export interface AccountInput {
  name: string;
  email: string;
  region: string;
  smtpHost: string;
  smtpPort: number;
  smtpUser: string;
  smtpPass: string;
  dailyLimit?: number;
  /** @nullable */
  signature?: string | null;
}

export interface AccountUpdate {
  name?: string;
  email?: string;
  region?: string;
  smtpHost?: string;
  smtpPort?: number;
  smtpUser?: string;
  smtpPass?: string;
  isActive?: boolean;
  dailyLimit?: number;
  /** @nullable */
  signature?: string | null;
}

export interface Template {
  id: number;
  name: string;
  subject: string;
  htmlContent: string;
  /** @nullable */
  textContent?: string | null;
  /** @nullable */
  category?: string | null;
  variables?: string[];
  createdAt: string;
  updatedAt?: string;
}

export interface TemplateInput {
  name: string;
  subject: string;
  htmlContent: string;
  textContent?: string;
  category?: string;
  variables?: string[];
}

export interface TemplateUpdate {
  name?: string;
  subject?: string;
  htmlContent?: string;
  textContent?: string;
  category?: string;
  variables?: string[];
}

export type AttachmentType = typeof AttachmentType[keyof typeof AttachmentType];


export const AttachmentType = {
  terms: 'terms',
  catalog: 'catalog',
  signature: 'signature',
  signature_banner: 'signature_banner',
} as const;

export interface Attachment {
  id: number;
  userId?: number;
  type?: AttachmentType;
  name: string;
  filename: string;
  mimeType: string;
  /** Base64 encoded file content */
  content?: string;
  isActive: boolean;
  assignedUserIds?: number[];
  createdAt: string;
}

export type AttachmentInputType = typeof AttachmentInputType[keyof typeof AttachmentInputType];


export const AttachmentInputType = {
  terms: 'terms',
  catalog: 'catalog',
  signature: 'signature',
  signature_banner: 'signature_banner',
} as const;

export interface AttachmentInput {
  userId?: number;
  type?: AttachmentInputType;
  name: string;
  filename: string;
  mimeType: string;
  /** Base64 encoded file content */
  content: string;
  isActive?: boolean;
  assignedUserIds?: number[];
}

export interface AttachmentUpdate {
  name?: string;
  /** Base64 encoded file content */
  content?: string;
  isActive?: boolean;
  assignedUserIds?: number[];
}

export interface SendEmailInput {
  accountId: number;
  templateId: number;
  /** Newline-separated email addresses or Name,email pairs */
  recipients: string;
  subject: string;
  /**
     * Override template HTML
     * @nullable
     */
  htmlContent?: string | null;
  /** Delay between sends in seconds */
  delaySeconds?: number;
}

export interface SendEmailResponse {
  queued: number;
  recipients: number;
  message?: string;
}

export type EmailHistoryStatus = typeof EmailHistoryStatus[keyof typeof EmailHistoryStatus];


export const EmailHistoryStatus = {
  sent: 'sent',
  failed: 'failed',
  pending: 'pending',
} as const;

export interface EmailHistory {
  id: number;
  accountId: number;
  accountName: string;
  recipientEmail: string;
  /** @nullable */
  recipientName?: string | null;
  subject: string;
  status: EmailHistoryStatus;
  /** @nullable */
  errorMessage?: string | null;
  /** @nullable */
  sentAt?: string | null;
  createdAt: string;
}

export interface EmailHistoryPage {
  data: EmailHistory[];
  total: number;
  page: number;
  limit: number;
}

export interface QueueStatus {
  pending: number;
  processing: number;
  total: number;
}

export interface DashboardStats {
  totalSent: number;
  totalFailed: number;
  totalPending: number;
  totalDelivered: number;
  openRate: number;
  bounceRate: number;
  totalAccounts: number;
  activeAccounts: number;
}

export interface AccountStat {
  accountId: number;
  accountName: string;
  region: string;
  sent: number;
  failed: number;
  pending: number;
  successRate?: number;
}

export interface DailyActivity {
  date: string;
  sent: number;
  failed: number;
}

export type EmailErrorErrorType = typeof EmailErrorErrorType[keyof typeof EmailErrorErrorType];


export const EmailErrorErrorType = {
  smtp_failed: 'smtp_failed',
  invalid_email: 'invalid_email',
  blocked: 'blocked',
  limit_reached: 'limit_reached',
  auth_error: 'auth_error',
  unknown: 'unknown',
} as const;

export interface EmailError {
  id: number;
  accountName: string;
  recipientEmail: string;
  errorType: EmailErrorErrorType;
  errorMessage: string;
  canRetry?: boolean;
  createdAt: string;
}

export type ListAttachmentsParams = {
userId?: number;
};

export type ListEmailHistoryParams = {
/**
 * @nullable
 */
accountId?: number | null;
/**
 * @nullable
 */
status?: ListEmailHistoryStatus;
/**
 * @nullable
 */
dateFrom?: string | null;
/**
 * @nullable
 */
dateTo?: string | null;
page?: number;
limit?: number;
};

export type ListEmailHistoryStatus = typeof ListEmailHistoryStatus[keyof typeof ListEmailHistoryStatus] | null;


export const ListEmailHistoryStatus = {
  sent: 'sent',
  failed: 'failed',
  pending: 'pending',
} as const;

export type GetDailyActivityParams = {
days?: number;
};

export type ListErrorsParams = {
limit?: number;
};