# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Email CRM API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /accounts:
    get:
      operationId: listAccounts
      tags: [accounts]
      summary: List sender accounts
      responses:
        "200":
          description: List of accounts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Account"
    post:
      operationId: createAccount
      tags: [accounts]
      summary: Create sender account (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AccountInput"
      responses:
        "201":
          description: Account created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
  /accounts/{id}:
    get:
      operationId: getAccount
      tags: [accounts]
      summary: Get account by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
    patch:
      operationId: updateAccount
      tags: [accounts]
      summary: Update account
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AccountUpdate"
      responses:
        "200":
          description: Updated account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
    delete:
      operationId: deleteAccount
      tags: [accounts]
      summary: Delete account
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /attachments:
    get:
      operationId: listAttachments
      tags: [attachments]
      summary: List attachments for a user (admin only)
      parameters:
        - name: userId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of attachments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Attachment"
    post:
      operationId: createAttachment
      tags: [attachments]
      summary: Upload an attachment for a user (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AttachmentInput"
      responses:
        "201":
          description: Attachment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Attachment"
  /attachments/{id}:
    patch:
      operationId: updateAttachment
      tags: [attachments]
      summary: Update attachment (toggle active, rename)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AttachmentUpdate"
      responses:
        "200":
          description: Updated attachment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Attachment"
    delete:
      operationId: deleteAttachment
      tags: [attachments]
      summary: Delete attachment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 1
    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
    User:
      type: object
      required: [id, email, name, role, createdAt]
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [admin, user]
        region:
          type: ["string", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
...
    Account:
      type: object
      required: [id, name, email, region, smtpHost, smtpPort, isActive, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        region:
          type: string
        smtpHost:
          type: string
        smtpPort:
          type: integer
        smtpUser:
          type: string
        smtpPass:
          type: string
        isActive:
          type: boolean
        dailyLimit:
          type: integer
        sentToday:
          type: integer
        signature:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    AccountInput:
      type: object
      required: [name, email, region, smtpHost, smtpPort, smtpUser, smtpPass]
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        region:
          type: string
        smtpHost:
          type: string
        smtpPort:
          type: integer
        smtpUser:
          type: string
        smtpPass:
          type: string
        dailyLimit:
          type: integer
        signature:
          type: ["string", "null"]
    AccountUpdate:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        region:
          type: string
        smtpHost:
          type: string
        smtpPort:
          type: integer
        smtpUser:
          type: string
        smtpPass:
          type: string
        isActive:
          type: boolean
        dailyLimit:
          type: integer
        signature:
          type: ["string", "null"]
...
    Attachment:
      type: object
      required: [id, name, filename, mimeType, isActive, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        type:
          type: string
          enum: [terms, catalog, signature, signature_banner]
        name:
          type: string
        filename:
          type: string
        mimeType:
          type: string
        content:
          type: string
          description: "Base64 encoded file content"
        isActive:
          type: boolean
        assignedUserIds:
          type: array
          items:
            type: integer
        createdAt:
          type: string
          format: date-time
    AttachmentInput:
      type: object
      required: [name, filename, mimeType, content]
      properties:
        userId:
          type: integer
        type:
          type: string
          enum: [terms, catalog, signature, signature_banner]
        name:
          type: string
        filename:
          type: string
        mimeType:
          type: string
        content:
          type: string
          description: "Base64 encoded file content"
        isActive:
          type: boolean
        assignedUserIds:
          type: array
          items:
            type: integer
    AttachmentUpdate:
      type: object
      properties:
        name:
          type: string
        content:
          type: string
          description: "Base64 encoded file content"
        isActive:
          type: boolean
        assignedUserIds:
          type: array
          items:
            type: integer
...