# lib/api-zod/src/generated/types/accountInput.ts
/**
 * Generated by orval v8.12.3 🍺
 * Do not edit manually.
 * Api
 * Email CRM API
 * OpenAPI spec version: 0.1.0
 */

export interface AccountInput {
  name: string;
  email: string;
  region: string;
  smtpHost: string;
  smtpPort: number;
  smtpUser: string;
  smtpPass: string;
  dailyLimit?: number;
  /** @nullable */
  signature?: string | null;
}